feat: change package name for better DX
This commit changes the main package to `k6-cli`, which is shorter and
easier for developers when installing; the ideal name would have been
`k6` but this has already been taken.

Consequently, the CI pipelines are updated due to changes in the names
of the release tarballs. As an added benefit, since the main package has
a different naming structure, it is now possible to use `*` to install
individual tarballs without conflict.